requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
charset-normalizer==2.1.1
click==8.0.4
dataclasses==0.8
Flask==2.0.3
frozenlist==1.2.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
pyee==9.0.4
slackclient==2.9.4
slackeventsapi==3.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
